+++ esa-20240910/alibabacloud_esa20240910/models/_list_certificates_by_record_response_body.py
@@ -161,6 +161,7 @@ class ListCertificatesByRecordResponseBodyResultCertificates(DaraModel):
         id: str = None,
         issuer: str = None,
         issuer_cn: str = None,
+        key_server_id: str = None,
         name: str = None,
         not_after: str = None,
         not_before: str = None,
@@ -187,6 +188,7 @@ class ListCertificatesByRecordResponseBodyResultCertificates(DaraModel):
         self.issuer = issuer
         # The Common Name of the certificate issuer.
         self.issuer_cn = issuer_cn
+        self.key_server_id = key_server_id
         # The certificate name.
         self.name = name
         # The time when the certificate expires.
@@ -238,6 +240,9 @@ class ListCertificatesByRecordResponseBodyResultCertificates(DaraModel):
 
         if self.issuer_cn is not None:
             result['IssuerCN'] = self.issuer_cn
+
+        if self.key_server_id is not None:
+            result['KeyServerId'] = self.key_server_id
 
         if self.name is not None:
             result['Name'] = self.name
@@ -296,6 +301,9 @@ class ListCertificatesByRecordResponseBodyResultCertificates(DaraModel):
 
         if m.get('IssuerCN') is not None:
             self.issuer_cn = m.get('IssuerCN')
+
+        if m.get('KeyServerId') is not None:
+            self.key_server_id = m.get('KeyServerId')
 
         if m.get('Name') is not None:
             self.name = m.get('Name')